+++ src/main/java/org/gradlex/maven/plugin/development/task/GenerateMavenPluginDescriptorTask.java
@@ -98,14 +98,16 @@ public abstract class GenerateMavenPluginDescriptorTask extends AbstractMavenPlu
             PluginToolsRequest pluginToolsRequest = createPluginToolsRequest(mavenProject, descriptor);
             // process upstream projects in order to scan base classes
             getUpstreamProjects().get().forEach(it -> {
-                File dir = it.getClassesDirs();
+                FileCollection dirs = it.getClassesDirs();
                 GAV gav = it.getGav();
-                DefaultArtifact artifact = new DefaultArtifact(
+                for (File dir : dirs) {
+                    DefaultArtifact artifact = new DefaultArtifact(
                             gav.getGroup(), gav.getArtifactId(), gav.getVersion(), "compile", "jar", null, new DefaultArtifactHandler()
                     );
                     artifact.setFile(dir);
                     pluginToolsRequest.getDependencies().add(artifact);
-                    mavenProject.addProjectReference(mavenProject(gav.getGroup(), gav.getArtifactId(), gav.getVersion(), it.getSourceDirectories(), classesDir));
+                }
+                mavenProject.addProjectReference(mavenProject(gav.getGroup(), gav.getArtifactId(), gav.getVersion(), it.getSourceDirectories(), classesDir));
             });
             populatePluginDescriptor(pluginToolsRequest);
         });