+++ src/main/java/org/gradlex/maven/plugin/development/task/UpstreamProjectDescriptor.java
@@ -16,6 +16,7 @@
 
 package org.gradlex.maven.plugin.development.task;
 
+import org.gradle.api.file.FileCollection;
 import org.gradle.api.tasks.CompileClasspath;
 import org.gradle.api.tasks.IgnoreEmptyDirectories;
 import org.gradle.api.tasks.InputFiles;
@@ -24,18 +25,16 @@ import org.gradle.api.tasks.PathSensitive;
 import org.gradle.api.tasks.PathSensitivity;
 import org.gradle.work.NormalizeLineEndings;
 
-import java.io.File;
-
 public final class UpstreamProjectDescriptor {
 
     private final GAV gav;
-    private final File classesDir;
-    private final File sourceDirectorie;
+    private final FileCollection classesDir;
+    private final FileCollection sourceDirectories;
 
-    public UpstreamProjectDescriptor(GAV gav, File classesDir, File sourcesDir) {
+    public UpstreamProjectDescriptor(GAV gav, FileCollection classesDir, FileCollection sourcesDir) {
         this.gav = gav;
         this.classesDir = classesDir;
-        this.sourceDirectorie = sourcesDir;
+        this.sourceDirectories = sourcesDir;
     }
 
     @Nested
@@ -45,7 +44,7 @@ public final class UpstreamProjectDescriptor {
 
     @CompileClasspath
     @InputFiles
-    public File getClassesDirs() {
+    public FileCollection getClassesDirs() {
         return classesDir;
     }
 
@@ -53,8 +52,8 @@ public final class UpstreamProjectDescriptor {
     @PathSensitive(PathSensitivity.RELATIVE)
     @IgnoreEmptyDirectories
     @NormalizeLineEndings
-    public File getSourceDirectories() {
-        return sourceDirectorie;
+    public FileCollection getSourceDirectories() {
+        return sourceDirectories;
     }
 }