+++ src/test/groovy/org/gradlex/maven/plugin/development/PluginDescriptorGenerationFuncTest.groovy
@@ -204,4 +204,15 @@ class PluginDescriptorGenerationFuncTest extends AbstractPluginFuncTest {
         !pluginDescriptor.hasDependency('com.google.guava:guava:28.0-jre')
     }
 
+    def "supports kotlin project dependency"() {
+        given:
+        multiProjectKotlinSetup()
+
+        when:
+        run(":plugin:generateMavenPluginDescriptor")
+
+        then:
+        pluginDescriptor("plugin/").hasDependency('org.example:kotlin-lib:1.0.0', '')
+    }
+
 }